perf(ipc): il hold probe — the read-lane hold becomes the §5.5.1 sync fast path's fourth leg
The 2026-08-02 il-anomalies charter (§2) counted the il randread
deficit's whole composition: kernel 231.7k vs il 211.0k IOPS on cold
rand-4k libaio qd16 because the kernel path's probe ladder (hot → hold
→ NVMe tier, before the ranged dispatch) serves ~294k ops/row from the
read-lane hold / hot tier at ~µs, while the il DIALED-P1.5 prelude
direct-drove those governor-DENIED misses device-true — ~220k missing
warm serves per row, both rows Little-close.

The fix (one site covers both law surfaces the charter named):
try_read_range_sync — the §5.5.1 sync serve engine every non-ddt ring
read crosses BEFORE the P1.5 ladder can submit device-true — gains the
hold as its fourth leg, in the handler ladder's order (hot strictly
first so warm hot entries keep funding the governor's payback basis).

- Foreign-thread discipline: the probe is serve_with_provenance —
  lock-free scc, no new locks, no blocking; binding currency is
  leg 3's own structural argument (inode read guard held, b_key from
  the CURRENT map). A miss counts and falls through unchanged
  (leg 3 → demote/direct-drive — fallback-is-correctness).
- R1b ledger visibility (the 3cd528b law's THIRD serve site): a
  demand-deposited entry's serve dispatches the admission ceremony
  (ghost touch → second-touch publish → protected hot re-landing) to
  the fuse3 per-core handler lanes via tpc_spawn — the 2026-07-26
  handoff-economy venue, never Handle::spawn from the foreign thread —
  exactly-once per serve; lane-fetch deposits stay ledger-invisible
  end to end (scan resistance stands).
- The sync hot leg now credits the hold copy (the handler hot arm's
  rule mirrored) so hold_evicted_unconsumed keeps meaning starvation
  on il rows.
- Engagement pair ipc_hold_probe_{serves,misses} on the stats inode;
  0 by construction under SQUEEZEFS_READ_LANE=0 (A0) and on
  direct_device_true mounts (kernel parity: ddt skips every warm leg
  by POLICY).

No new lock-free protocol (the hold's scc/atomic machinery is
unchanged — no loom owed per the house mandate). All 6 red contracts
green; read family + ipc family + parity/saturation/ranged suites
green at --test-threads=1; clippy -D warnings + fmt clean.

+++ tests/ipc_hold_probe_tests.rs
@@ -328,11 +328,7 @@ impl Fixture {
     /// a cold attr cache before it can probe anything — warm it through
     /// the FUSE getattr handler (its miss arm seeds the cache).
     async fn warm_attr(&self, ino: u64) {
-        let _ = self
-            .fs
-            .getattr(req(), ino, None, 0)
-            .await
-            .expect("getattr");
+        let _ = self.fs.getattr(req(), ino, None, 0).await.expect("getattr");
         assert!(
             self.fs.attr_cache.get(&ino).is_some(),
             "fixture: getattr must seed the attr cache"
@@ -557,7 +553,10 @@ async fn hold_probe_serves_bytes_exact_and_ceremony_lands_exactly_once() {
         fx.fs.router.cache.read_lane_hold.contains(&k0),
         "fixture: the demand fill must deposit in the hold"
     );
-    assert!(!fx.tier_has(&k0), "fixture: first touch skipped the publish");
+    assert!(
+        !fx.tier_has(&k0),
+        "fixture: first touch skipped the publish"
+    );
     // Kill the hot copy: the ring read's ONLY warm source is the hold.
     fx.fs.router.cache.hot_block.remove(&k0);
     fx.warm_attr(ino).await;
@@ -739,9 +738,7 @@ async fn hold_probe_miss_counts_and_demote_path_is_preserved() {
 
     let misses0 = METRICS.ipc_hold_probe_misses.load(Ordering::Relaxed);
     let serves0 = METRICS.ipc_hold_probe_serves.load(Ordering::Relaxed);
-    let demote0 = METRICS
-        .ipc_fast_path_miss_demotions
-        .load(Ordering::Relaxed);
+    let demote0 = METRICS.ipc_fast_path_miss_demotions.load(Ordering::Relaxed);
 
     let r = tokio::task::block_in_place(|| session.ring_pread_spin(binding, 4096, 4096));
     assert_eq!(r, 4096, "the miss must still serve through the handoff");
@@ -759,10 +756,7 @@ async fn hold_probe_miss_counts_and_demote_path_is_preserved() {
         0
     );
     assert_eq!(
-        METRICS
-            .ipc_fast_path_miss_demotions
-            .load(Ordering::Relaxed)
-            - demote0,
+        METRICS.ipc_fast_path_miss_demotions.load(Ordering::Relaxed) - demote0,
         1,
         "the miss demotes to the async handoff exactly as before the probe"
     );
@@ -875,7 +869,7 @@ async fn sync_hot_leg_serves_credit_the_hold_toward_retirement() {
 
     // Full-coverage ring read of block 0 through the HOT leg (probe
     // order: hot strictly before hold — these must be hot serves).
-    let steps = (BS / 4096) as u64;
+    let steps = BS / 4096;
     tokio::task::block_in_place(|| {
         for i in 0..steps {
             let r = session.ring_pread_spin(binding, i * 4096, 4096);
